+++ docs/source/release_notes.rst
@@ -73,6 +73,16 @@ Availability: `AutomationHub`_, `Galaxy`_, `GitHub`_
   and 'virtual-storage-resources' returned by the module will be empty arrays
   for non-FCP storage groups. (e.g. NVMe). (issue #864)
 
+* Fixed that on HMC versions 2.14 and 2.15, the zhmc_adapter_list module
+  failed because it tried to use the "List Permitted Adapters" operation
+  that was added in HMC version 2.16 (actually API version 4.10).
+  (issue #850)
+
+* Fixed that on HMC versions 2.14 and 2.15, the zhmc_partition_list module
+  failed because it tried to use the 'additional-properties' query parameter
+  that was added in HMC version 2.16 (actually API version 4.10).
+  (issue #850)
+
 **Enhancements:**
 
 * Added support for Python 3.12. (issue #796)
@@ -98,6 +108,19 @@ Availability: `AutomationHub`_, `Galaxy`_, `GitHub`_
   - packaging to 21.3 (on Python >= 3.6)
   - PyYAML to 6.0.1 (on Python >= 3.6)
   - jsonschema to 4.10.0 (on Python >= 3.7)
+
+* In the 'zhmc_adapter_list' module, improved the use of the "Permitted
+  Adapters" operation so that it is now also used when the 'additional_properties'
+  module parameter is used and the HMC API version is 4.10 or higher.
+  (related to issue #850)
+
+* Docs: In the 'zhmc_lpar_list' module, clarified that the use of the "List
+  Permitted Logical Partitions" operation does not affect the module result
+  data. (related to issue #850)
+
+* Docs: In the 'zhmc_partition_list' module, clarified that the use of the "List
+  Permitted Partitions" operation does not affect the module result data.
+  (related to issue #850)
 
 **Cleanup:**